Split off tests in `schema` module
